+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Capture the options of the failing validation rule for more accurate error messages.
 - Added `Iterator::toArray()` as a shortcut to `iterator_to_array`
 - `Query::where()` now accepts models as conditions
+- Doctrine 3 compatibility if using the DBAL driver
 
 ### Changed
 - Moved adding event listeners and dispatching events to `EventManager`